requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# LangChain & LangGraph
langchain>=0.1.6
langchain-groq>=0.0.1
langchain-core>=0.1.19
langgraph>=0.0.26

# Groq API
groq>=0.4.2

# OCR
paddleocr>=2.7.0
paddlepaddle>=3.0.0
Pillow>=10.2.0
pdf2image>=1.17.0
pypdf>=4.0.0

# Data Validation
pydantic>=2.6.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.1
httpx>=0.26.0

# Logging & Monitoring
loguru>=0.7.2

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.4
